Seperate src into common, node. and browser folders
