+++ client/src/common/virtualFiles.ts
