+++ client/tsconfig.browser.json
@@ -1,0 +1,16 @@
+{
+  "extends": "./tsconfig.base.json",
+  "compilerOptions": {
+    "lib": ["ES6", "ES2020", "WebWorker"],
+    "paths": {
+      "path": ["path-browserify"]
+    }
+  },
+  "exclude": [
+    "node_modules",
+    "src/test-runner",
+    "src/test",
+    "src/extension.ts"
+  ],
+  "include": ["src/common", "src/browser"]
+}
