+++ tsconfig.json
@@ -1,7 +1,9 @@
 {
   "files": [],
   "references": [
-    { "path": "./client/tsconfig.build.json" },
-    { "path": "./server/tsconfig.build.json" }
+    { "path": "./client/tsconfig.node.json" },
+    { "path": "./client/tsconfig.browser.json" },
+    { "path": "./server/tsconfig.node.json" },
+    { "path": "./server/tsconfig.browser.json" }
   ]
 }
